rawger/Reader: Refactor to decouple Reader and Paginator view models.
Splits the ReaderViewModel into ReaderViewModel and ReaderPaginatorViewModel. This fixes the previous commit's problem where the ReaderViewModel is redundantly declared in two views.Also lays the foundation for the Paginator overhaul.
